third_party/blink/web_tests/virtual/html-anchor-attribute-disabled/external/wpt/html/semantics/popovers/popover-anchor-idl-property.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] popover anchorElement IDL property returns the anchor element
  assert_equals: expected (object) Element node <button id="b1">This is an anchor button</button> but got (undefined) undefined
[FAIL] popover anchorElement is settable
  assert_equals: expected (object) Element node <button id="b1">This is an anchor button</button> but got (undefined) undefined
[FAIL] anchorElement affects anchor positioning
  assert_equals: expected (object) null but got (undefined) undefined
Harness: the test ran to completion.
